Finalize recipe and add new image
The fried potatoes page was finalized.

A new image was added to the page.

Below the image there are two paragraphs containing the description of
the recipe.

An unordered and an ordered list is at the bottom of the page: the first
one enumerates all the ingredients needed for the recipe, the last one
shows all the steps needed to make the dish.